requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.2
gpg==1.10.0
GDAL==3.11.0
geopandas==1.1.0
h3==4.2.2
h3pandas==0.3.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.3
networkx==3.5
numpy==2.3.0
osmnx==2.0.3
packaging==25.0
pandas==2.3.0
pillow==11.2.1
pyarrow==20.0.0
pyogrio==0.11.0
pyparsing==3.2.3
pyproj==3.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
raster2xyz==0.1.3
requests==2.32.3
shapely==2.1.1
six==1.17.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
wheel==0.45.1